Pull the master branch without --ff-only
The --ff-only pull will succeed even if it couldn't fast forward. In
that case, the current branch is behind the origin, and this will result
in colossal failure of `mob continue` on the other side which *is*
up-to-date with the origin. This is due to the fact that we do a
`git reset` when doing `mob continue`, which resets the pushed branch to
the current main branch. That will undo all changes in the up-to-date
main branch that the non-up-to-date branch was behind...

using git merge might result in a merge conflict, but it will halt the
process at that point, so that the user can intervene and fix the
problem first.