Always push the local branch before creating a wip commit
If the person doing `mob continue` doesn't have the latest commits to
apply the wip changes to, it can result in a horrable git diff fail, or
worse, a rebase problem.

This sometimes occurred when we did a `mob switch` without pushing our
committed changes first. This commit changes this, so that `mob switch`
pushes our local commits before creating a wip commit on the wip branch.